+++ src/main/java/com/kodilla/ecommercee/mapper/GroupMapper.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.mapper;
+
+public class GroupMapper {
+}
